requirements.txt
streamlit>=1.28.0
openai>=1.12.0
anthropic>=0.25.0
notion-client>=2.2.1
pydub>=0.25.1
audioop-lts>=0.2.1; python_version >= "3.13"
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
